Namespace the webmoney decorator
