+++ spec/models/spree/web_money_decorator_spec.rb
@@ -1,10 +1,10 @@
 require "spec_helper"
 
-describe WebMoneyDecorator, type: :model do
+describe Spree::WebMoneyDecorator, type: :model do
   context 'when source is webmoney' do
     it 'delegates methods' do
       source = Spree::WebMoney.new(prepaid_cards: [], short_amount: 1000, payment_uuid: "123")
-      decorator = WebMoneyDecorator.new(source)
+      decorator = Spree::WebMoneyDecorator.new(source)
 
       expect(decorator.prepaid_cards).to eq([])
       expect(decorator.payment_uuid).to eq("123")
@@ -14,7 +14,7 @@ describe WebMoneyDecorator, type: :model do
 
   context 'when source is nill' do
     it 'delegates methods to nil' do
-      decorator = WebMoneyDecorator.new(nil) 
+      decorator = Spree::WebMoneyDecorator.new(nil) 
 
       expect(decorator.prepaid_cards).to be_nil
       expect(decorator.payment_uuid).to be_nil
